got the TI version mostly building in sdcc
Problems are:
- some L-suffixed hex constants in MIRACL SHA256 code that sdcc doesn't like for some reason. Removing the L made a warning go away, but I don't know if the value got through properly in this case.
- putc() isn't defined
- hasn't been run at all!
- needs a few more transformations, from sdcc's hex format to bin to packaged app

I am thinking about ditching sdcc for z88dk, as they seem to more directly support the calculators.  If I stay with sdcc, I'll be on my own for packaging the app and writing some glue to call in to the calculator's I/O routines.  Did I already try z88dk and decide it wouldn't work for some reason?